requirements.txt
fastapi
uvicorn[standard]
pydantic
pytest
openenv-core
requests
gradio
huggingface-hub
